requirements.txt
streamlit==1.27.0
tensorflow==2.14.0
numpy==1.26.2
pillow==10.0.0
